face_recognizer/requirements.txt
face_recognition
scipy
flask
werkzeug
Pillow